# Assorted optimizations