@sergi 2018-01-25
Hi @josharian,

Thanks so much for this PR, this is great work. My apologies for getting back to you so late, but better late than never!